# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "本地化的 Bangumi 数据服务。利用 Archive 每周导出的 wiki 数据导入本地 SQLite，\n提供条目/人物/角色/章节等 REST 查询接口与单二进制管理接口。\n统一响应包装：成功 {\"ok\": true, \"data\": ...}；失败 {\"ok\": false, \"error\": \"...\"}。\n数据接口均为只读；管理接口（/api/admin/*）需要 X-Admin-Token 请求头（未配置 token 时放行）。\n交互式文档：/docs 或 /swagger/index.html；本文档由 swag 从代码注解自动生成（make docs）。",
        "title": "Bangumi 本地数据搜索 API",
        "contact": {
            "name": "GitHub",
            "url": "https://github.com/nz3u/bangumi-data"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/characters/search": {
            "get": {
                "description": "q 匹配原名与 infobox 简体中文名（≥3 字符走全文索引，更短回退 LIKE 子串匹配）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "角色搜索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "角色类型：1 角色 / 2 机体 / 3 组织",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.characterSearchData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/characters/{id}": {
            "get": {
                "description": "含基本信息、infobox（结构化字段）、简介、出演作品、声优/演员（按人物去重）与角色关联（双向）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "角色详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.characterDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.apiEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "api.characterBrief": {
            "type": "object",
            "properties": {
                "collects": {
                    "type": "integer"
                },
                "comments": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "name_cn": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "api.characterDetail": {
            "type": "object",
            "properties": {
                "collects": {
                    "type": "integer"
                },
                "comments": {
                    "type": "integer"
                },
                "cvs": {
                    "description": "声优/演员（person_characters 关联）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.cvItem"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "infobox": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wiki.Field"
                    }
                },
                "name": {
                    "type": "string"
                },
                "name_cn": {
                    "type": "string"
                },
                "relations": {
                    "description": "角色关联（person_relations 中 crt 行，双向）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.personRelationItem"
                    }
                },
                "role": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                },
                "subjects": {
                    "description": "出演作品",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.characterSubjectItem"
                    }
                },
                "summary": {
                    "type": "string"
                }
            }
        },
...
        "api.characterSearchData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.characterBrief"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.characterSubjectItem": {
            "type": "object",
            "properties": {
                "char_type": {
                    "type": "integer"
                },
                "char_type_name": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_cn": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "subject_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "type_name": {
                    "type": "string"
                }
            }
        },
...
        "api.cvItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "person_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                },
                "subject_name": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "type_name": {
                    "type": "string"
                }
            }
        },
...
        "api.personRelationItem": {
            "type": "object",
            "properties": {
                "ended": {
                    "type": "boolean"
                },
                "person_id": {
                    "type": "integer"
                },
                "person_type": {
                    "type": "string"
                },
                "related_name": {
                    "type": "string"
                },
                "related_person_id": {
                    "type": "integer"
                },
                "related_type": {
                    "type": "integer"
                },
                "relation_name": {
                    "type": "string"
                },
                "relation_type": {
                    "type": "integer"
                },
                "spoiler": {
                    "type": "boolean"
                }
            }
        },
...
        "wiki.Field": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        }
    }
}